chore(tests): promote Playwright experimental tests to stable and clean up superseded Cypress tests
Move chart, dashboard, and dataset Playwright E2E tests from experimental/
to stable test directories so they run as required CI blockers. Delete the
Cypress chart_list and dataset_list tests that are now fully replaced, and
remove orphaned helpers from Cypress utils.
