+++ superset-frontend/cypress-base/cypress/e2e/explore/utils.ts
@@ -23,18 +23,6 @@ export function interceptFiltering() {
   cy.intercept('GET', `**/api/v1/chart/?q=*`).as('filtering');
 }
 
-export function interceptBulkDelete() {
-  cy.intercept('DELETE', `**/api/v1/chart/?q=*`).as('bulkDelete');
-}
-
-export function interceptDelete() {
-  cy.intercept('DELETE', `**/api/v1/chart/*`).as('delete');
-}
-
-export function interceptFavoriteStatus() {
-  cy.intercept('GET', '**/api/v1/chart/favorite_status/*').as('favoriteStatus');
-}
-
 export function interceptUpdate() {
   cy.intercept('PUT', `**/api/v1/chart/*`).as('update');
 }
@@ -56,17 +44,6 @@ export function interceptExploreGet() {
     method: 'GET',
     url: /.*\/api\/v1\/explore\/\?(form_data_key|dashboard_page_id|slice_id)=.*/,
   }).as('getExplore');
-}
-
-export function setFilter(filter: string, option: string) {
-  interceptFiltering();
-
-  cy.get(`[aria-label^="${filter}"]`).first().click();
-  cy.get(`.ant-select-item-option[title="${option}"]`).first().click({
-    force: true,
-  });
-
-  cy.wait('@filtering');
 }
 
 export function saveChartToDashboard(chartName: string, dashboardName: string) {
