+++ superset-frontend/playwright/tests/chart/chart-test-helpers.ts
@@ -18,45 +18,59 @@
  */
 
 import type { Page, TestInfo } from '@playwright/test';
-import type { TestAssets } from '../../../helpers/fixtures/testAssets';
-import { apiPostDashboard } from '../../../helpers/api/dashboard';
+import type { TestAssets } from '../../helpers/fixtures/testAssets';
+import { apiPostChart } from '../../helpers/api/chart';
+import { getDatasetByName } from '../../helpers/api/dataset';
 
-interface TestDashboardResult {
+interface TestChartResult {
   id: number;
   name: string;
 }
 
-interface CreateTestDashboardOptions {
-  /** Prefix for generated name (default: 'test_dashboard') */
+interface CreateTestChartOptions {
+  /** Prefix for generated name (default: 'test_chart') */
   prefix?: string;
 }
 
 /**
- * Creates a test dashboard via the API for E2E testing.
+ * Creates a test chart via the API for E2E testing.
+ * Uses the members_channels_2 dataset (loaded via --load-examples).
  *
  * @example
- * const { id, name } = await createTestDashboard(page, testAssets, test.info());
+ * const { id, name } = await createTestChart(page, testAssets, test.info());
  *
  * @example
- * const { id, name } = await createTestDashboard(page, testAssets, test.info(), {
+ * const { id, name } = await createTestChart(page, testAssets, test.info(), {
  *   prefix: 'test_delete',
  * });
  */
-export async function createTestDashboard(
+export async function createTestChart(
   page: Page,
   testAssets: TestAssets,
   testInfo: TestInfo,
-  options?: CreateTestDashboardOptions,
-): Promise<TestDashboardResult> {
-  const prefix = options?.prefix ?? 'test_dashboard';
+  options?: CreateTestChartOptions,
+): Promise<TestChartResult> {
+  const prefix = options?.prefix ?? 'test_chart';
   const name = `${prefix}_${Date.now()}_${testInfo.parallelIndex}`;
 
-  const response = await apiPostDashboard(page, {
-    dashboard_title: name,
+  // Look up the members_channels_2 dataset for chart creation
+  const dataset = await getDatasetByName(page, 'members_channels_2');
+  if (!dataset) {
+    throw new Error(
+      'members_channels_2 dataset not found — run Superset with --load-examples',
+    );
+  }
+
+  const response = await apiPostChart(page, {
+    slice_name: name,
+    datasource_id: dataset.id,
+    datasource_type: 'table',
+    viz_type: 'table',
+    params: '{}',
   });
 
   if (!response.ok()) {
-    throw new Error(`Failed to create test dashboard: ${response.status()}`);
+    throw new Error(`Failed to create test chart: ${response.status()}`);
   }
 
   const body = await response.json();
@@ -64,11 +78,11 @@ export async function createTestDashboard(
   const id = body.result?.id ?? body.id;
   if (!id) {
     throw new Error(
-      `Dashboard creation returned no id. Response: ${JSON.stringify(body)}`,
+      `Chart creation returned no id. Response: ${JSON.stringify(body)}`,
     );
   }
 
-  testAssets.trackDashboard(id);
+  testAssets.trackChart(id);
 
   return { id, name };
 }
