+++ superset-frontend/playwright/tests/dashboard/dashboard-list.spec.ts
@@ -20,25 +20,25 @@
 import {
   test as testWithAssets,
   expect,
-} from '../../../helpers/fixtures/testAssets';
-import { DashboardListPage } from '../../../pages/DashboardListPage';
-import { DeleteConfirmationModal } from '../../../components/modals/DeleteConfirmationModal';
-import { ImportDatasetModal } from '../../../components/modals/ImportDatasetModal';
-import { Toast } from '../../../components/core/Toast';
+} from '../../helpers/fixtures/testAssets';
+import { DashboardListPage } from '../../pages/DashboardListPage';
+import { DeleteConfirmationModal } from '../../components/modals/DeleteConfirmationModal';
+import { ImportDatasetModal } from '../../components/modals/ImportDatasetModal';
+import { Toast } from '../../components/core/Toast';
 import {
   apiGetDashboard,
   apiDeleteDashboard,
   apiExportDashboards,
   getDashboardByName,
   ENDPOINTS,
-} from '../../../helpers/api/dashboard';
+} from '../../helpers/api/dashboard';
 import { createTestDashboard } from './dashboard-test-helpers';
-import { waitForGet, waitForPost } from '../../../helpers/api/intercepts';
+import { waitForGet, waitForPost } from '../../helpers/api/intercepts';
 import {
   expectStatusOneOf,
   expectValidExportZip,
-} from '../../../helpers/api/assertions';
-import { TIMEOUT } from '../../../utils/constants';
+} from '../../helpers/api/assertions';
+import { TIMEOUT } from '../../utils/constants';
 
 /**
  * Extend testWithAssets with dashboardListPage navigation (beforeEach equivalent).
