+++ superset-frontend/playwright/tests/dataset/dataset-test-helpers.ts
@@ -18,8 +18,8 @@
  */
 
 import type { Page, TestInfo } from '@playwright/test';
-import type { TestAssets } from '../../../helpers/fixtures/testAssets';
-import { createTestVirtualDataset } from '../../../helpers/api/dataset';
+import type { TestAssets } from '../../helpers/fixtures/testAssets';
+import { createTestVirtualDataset } from '../../helpers/api/dataset';
 
 interface TestDatasetResult {
   id: number;
